+++ src/components/FeedbackBox/FeedbackBox.tsx
@@ -182,7 +182,6 @@ export const FeedbackBox: ChakraComponent<
         isFormView && notificationText ? (
           <Notification
             isCentered
-            noMargin
             notificationContent={notificationText}
             showIcon={false}
             p="0"