+++ src/theme/components/button.ts
@@ -131,32 +131,6 @@ export const secondary = definePartsStyle(({ buttonSize = "medium" }: any) => ({
     },
   },
 }));
-// The "link" type is deprecated but we still want to style
-// it correctly for backwards compatibility.
-const link = definePartsStyle(({ buttonSize = "medium" }: any) => ({
-  base: {
-    bg: "transparent",
-    color: "ui.link.primary",
-    textDecoration: "underline",
-    ...generalSizeValues(buttonSize),
-    _disabled: {
-      bg: "transparent",
-    },
-    _hover: {
-      bg: "transparent",
-      color: "ui.link.secondary",
-    },
-    _dark: {
-      color: "dark.ui.link.primary",
-      _disabled: {
-        bg: "transparent",
-      },
-      _hover: {
-        color: "dark.ui.link.secondary",
-      },
-    },
-  },
-}));
 const text = definePartsStyle(({ buttonSize = "medium" }: any) => ({
   base: {
     bg: "transparent",
@@ -273,8 +247,6 @@ const Button = defineMultiStyleConfig({
     primary,
     secondary,
     text,
-    // Deprecated type. Use "text" instead.
-    link,
     pill,
     iconOnly,
     callout,