+++ src/theme/components/global.ts
@@ -27,7 +27,7 @@ const activeFocus = (darkMode = false) => ({
 // Used in `Checkbox` and `Radio`.
 const checkboxRadioLabelStyles = {
   userSelect: "none",
-  fontWeight: "text.default",
+  fontWeight: "body.default",
   marginBottom: "0",
   marginStart: "xs",
   width: "100%",
@@ -129,6 +129,7 @@ const labelLegendText = {
 const labelLegendTextSpecialSpacing = {
   marginBottom: "xxxs",
 };
+
 // Used for the disabled state of Select and TextInput
 const selectTextInputDisabledStyles = {
   bg: "ui.disabled.secondary",